requirements.txt

bcrypt==4.0.1
cryptography==41.0.3
dnspython==2.4.1
fastapi==0.75.0
Jinja2==3.1.2
passlib==1.7.4
pydantic==1.9.0
pymongo==4.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.6
PyYAML==6.0
rsa==4.9
starlette==0.17.1
typing_extensions==4.7.1
uvicorn==0.17.6
python-dotenv